Added mason and ocamlls

+++ dotfiles_setup.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 d="$(pwd)/${0}"
 root_dir=${d%/*}
-ln -s  $root_dir $HOME/.config/${root_dir##*/}
+[ -L $HOME/.config/${root_dir##*/} ] || ln -s  $root_dir $HOME/.config/${root_dir##*/}
